# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chowdhary Mart API — hyperlocal e-commerce platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List nearby stores
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
  /stores/{storeId}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store by ID
      parameters:
        - $ref: "#/components/parameters/StoreIdParam"
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Products ──────────────────────────────────────────────────────────────
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List / search products
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: storeId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 40
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /vendor/products:
    get:
      operationId: listVendorProducts
      tags: [products]
      summary: List vendor's own products
      responses:
        "200":
          description: Vendor products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (vendor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /vendor/products/{productId}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (vendor)
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (vendor)
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Addresses ─────────────────────────────────────────────────────────────
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get my wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Orders ────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /vendor/store:
    get:
      operationId: getVendorStore
      tags: [stores]
      summary: Get vendor's own store
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
    patch:
      operationId: updateVendorStore
      tags: [stores]
      summary: Update vendor store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
  /vendor/orders:
    get:
      operationId: listVendorOrders
      tags: [orders]
      summary: List vendor's incoming orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /vendor/orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (vendor)
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  parameters:
    StoreIdParam:
      name: storeId
      in: path
      required: true
      schema:
        type: integer
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: integer
    OrderIdParam:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ───────────────────────────────────────────────────────────────
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        colorClass:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: integer

    # ── Store ──────────────────────────────────────────────────────────────
    Store:
      type: object
      required: [id, name, lat, lng, address]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: ["string", "null"]
        rating:
          type: ["string", "null"]
        ratingCount:
          type: ["integer", "null"]
        isOpen:
          type: boolean
        estimatedDeliveryMins:
          type: ["integer", "null"]
        deliveryFee:
          type: ["string", "null"]
        freeDeliveryAbove:
          type: ["string", "null"]
        minOrderValue:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
    StoreDetail:
      allOf:
        - $ref: "#/components/schemas/Store"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            phone:
              type: ["string", "null"]
            isVerified:
              type: boolean
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        isOpen:
          type: boolean
        phone:
          type: string
        estimatedDeliveryMins:
          type: integer

    # ── Product ────────────────────────────────────────────────────────────
    Product:
      type: object
      required: [id, name, price, mrp]
      properties:
        id:
          type: integer
        storeId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        mrp:
          type: string
        discountPercent:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        weight:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        stock:
          type: integer
        rating:
          type: ["string", "null"]
        reviewCount:
          type: ["integer", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            specifications:
              type: object
              additionalProperties:
                type: string
            tags:
              type: array
              items:
                type: string
            category:
              $ref: "#/components/schemas/Category"
            store:
              $ref: "#/components/schemas/Store"
    ProductListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        hasMore:
          type: boolean
    ProductInput:
      type: object
      required: [name, price, mrp, stock]
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        brandId:
          type: integer
        price:
          type: string
        mrp:
          type: string
        images:
          type: array
          items:
            type: string
        weight:
          type: string
        unit:
          type: string
        stock:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        mrp:
          type: string
        stock:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        images:
          type: array
          items:
            type: string

    # ── Address ────────────────────────────────────────────────────────────
...
    WishlistItem:
      type: object
      required: [id, productId, addedAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        addedAt:
          type: string
        product:
          $ref: "#/components/schemas/Product"
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer

    # ── Order ──────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        storeId:
          type: integer
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: string
        deliveryFee:
          type: string
        discount:
          type: string
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: ["string", "null"]
        walletUsed:
          type: ["string", "null"]
        total:
          type: string
        loyaltyPointsEarned:
          type: ["integer", "null"]
        estimatedDeliveryMins:
          type: ["integer", "null"]
        createdAt:
          type: string
        store:
          $ref: "#/components/schemas/Store"
...
    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string
          enum: [cod, online, wallet, upi]
        couponCode:
          type: string
        useWallet:
          type: boolean
        notes:
          type: string
...
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ── Tracking ───────────────────────────────────────────────────────────
...